+++ src/main/java/org/codinjutsu/tools/jenkins/view/action/SelectViewAction.java
@@ -28,7 +28,6 @@ import com.intellij.ui.components.JBList;
 import org.codinjutsu.tools.jenkins.model.FavoriteView;
 import org.codinjutsu.tools.jenkins.model.View;
 import org.codinjutsu.tools.jenkins.view.BrowserPanel;
-import org.codinjutsu.tools.jenkins.view.JenkinsNestedViewComboRenderer;
 import org.codinjutsu.tools.jenkins.view.JenkinsViewComboRenderer;
 import org.jetbrains.annotations.NotNull;
 
@@ -76,12 +75,7 @@ public class SelectViewAction extends DumbAwareAction implements CustomComponent
         }
 
         final JBList<View> viewList = new JBList<>(unflattenViews);
-
-        if (hasNestedViews(unflattenViews)) {
-            viewList.setCellRenderer(new JenkinsNestedViewComboRenderer());
-        } else {
-            viewList.setCellRenderer(new JenkinsViewComboRenderer());
-        }
+        viewList.setCellRenderer(new JenkinsViewComboRenderer());
         return viewList;
     }
 
@@ -119,14 +113,6 @@ public class SelectViewAction extends DumbAwareAction implements CustomComponent
         }
 
         return flattenViewList;
-    }
-
-
-    private static boolean hasNestedViews(List<View> views) {
-        for (View view : views) {
-            if (view.hasNestedView()) return true;
-        }
-        return false;
     }
 
     private class MyMouseAdapter extends MouseAdapter {